Legacy code is outside of the main repo now (i.e. exclude it from coverage ).  Moved out legacy code in mapping.py
